Merge CHERI code into this repo
Eventually we're going to want to have CHERI support in the upstream sail-riscv repo, so I think merging it here makes sense. This removes the need to deal with multiple submodules, and in future it should simplify the build system and generally be easier to maintain.

I haven't updated the CMake yet, so you still need to use the Makefile to build it.

This code is exactly as in the CHERI alliance sail-cheri-riscv repo commit 245f608735fc4fbeced497203b8114fd3127dec2.